Move admin helper functions out of routes
